Pred sortiranje pogledamo ce je slucajno seznam ze sortiran, drugace minor fixes

+++ c-single-threaded/array-list-int.h
@@ -4,8 +4,8 @@
 
 typedef struct __ArrayListInt {
 	int *arr;
-	long elements;
-	long size;
+	unsigned int elements;
+	unsigned int size;
 } ArrayListInt;
